@@ -1,47 +1,47 @@
 {
-  "name": "angular",
-  "version": "1.0.0",
-  "private": true,
+  "name": "adminlte-angular",
+  "version": "0.0.0",
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
     "build": "ng build",
-    "test": "ng test",
-    "lint": "ng lint",
-    "e2e": "ng e2e"
+    "watch": "ng build --watch --configuration development",
+    "test": "ng test"
   },
+  "private": true,
   "dependencies": {
-    "@angular/common": "8.2.14",
-    "@angular/compiler": "8.2.14",
-    "@angular/core": "8.2.14",
-    "@angular/fire": "5.3.0",
-    "@angular/forms": "8.2.14",
-    "@angular/http": "7.2.16",
-    "@angular/platform-browser": "8.2.14",
-    "@angular/platform-browser-dynamic": "8.2.14",
-    "@angular/router": "8.2.14",
-    "@exalif/ngx-breadcrumbs": "7.1.1",
-    "@firebase/app": "0.5.2",
-    "@ng-bootstrap/ng-bootstrap": "5.2.1",
-    "@ng-select/ng-select": "3.7.1",
-    "@ngrx/store": "8.6.0",
-    "@swimlane/ngx-datatable": "16.0.3",
-    "@types/node": "13.5.0",
-    "angularfire2": "5.3.0",
-    "bootstrap": "4.4.1",
-    "chart.js": "2.9.3",
-    "firebase": "7.7.0",
-    "ionicons": "4.6.3",
-    "jquery": "3.4.1",
-    "lodash": "4.17.15",
-    "lodash.templatesettings": "4.2.0",
-    "ng2-ace-editor": "0.3.9",
-    "ng2-charts": "2.3.0",
-    "popper.js": "1.16.1",
-    "rxjs": "6.5.4",
-    "rxjs-compat": "6.5.4",
-    "zone.js": "0.10.2",
-    "core-js": "2.5.7"
+    "@angular/animations": "^15.2.0",
+    "@angular/common": "^15.2.0",
+    "@angular/compiler": "^15.2.0",
+    "@angular/core": "^15.2.0",
+    "@angular/fire": "^7.6.1",
+    "@angular/forms": "^15.2.0",
+    "@angular/platform-browser": "^15.2.0",
+    "@angular/platform-browser-dynamic": "^15.2.0",
+    "@angular/router": "^15.2.0",
+    "@exalif/ngx-breadcrumbs": "^16.0.0",
+    "@swimlane/ngx-datatable": "^20.1.0",
+    "bootstrap": "^5.3.1",
+    "chart.js": "^4.4.0",
+    "ionicons": "^7.1.2",
+    "jquery": "^3.7.1",
+    "lodash": "^4.17.21",
+    "popper.js": "^1.16.1",
+    "rxjs": "~7.8.0",
+    "tslib": "^2.3.0",
+    "zone.js": "~0.12.0"
   },
-  "devDependencies": {}
+  "devDependencies": {
+    "@angular-devkit/build-angular": "^15.2.2",
+    "@angular/cli": "~15.2.2",
+    "@angular/compiler-cli": "^15.2.0",
+    "@types/jasmine": "~4.3.0",
+    "jasmine-core": "~4.5.0",
+    "karma": "~6.4.0",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage": "~2.2.0",
+    "karma-jasmine": "~5.1.0",
+    "karma-jasmine-html-reporter": "~2.0.0",
+    "typescript": "~4.9.4"
+  }
 }
