@@ -1,16 +1,17 @@
 import { NgModule } from '@angular/core';
-import { BrowserModule } from '@angular/platform-browser';
 import { Routes, RouterModule, PreloadAllModules } from '@angular/router';
-const routes: Routes = [  
+const routes: Routes = [
   {
     path: '',
-    loadChildren: './layout/admin/admin-layout.module#AdminLayoutModule',    
+    loadChildren: () => import('./layout/admin/admin-layout.module').then(m => m.AdminLayoutModule)
   },
   {
     path: 'user',
-    loadChildren: './layout/default/default-layout.module#DefaultLayoutModule',
+    loadChildren: () => import('./layout/default/default-layout.module').then(m => m.DefaultLayoutModule)
   },
-  { path: '**', redirectTo: '' }
+  {
+    path: '**', redirectTo: ''
+  }
 ];
 
 @NgModule({
