@@ -10,13 +10,16 @@ const routes: Routes = [
     component: DefaultLayoutComponent,
     children: [
       {
-        path: '', loadChildren: '../../modules/default/default.module#DefaultModule',
+        path: '',  
+        loadChildren: () => import('../../modules/default/default.module').then(m => m.DefaultModule)
       },
       {
-        path: 'login', loadChildren: '../../modules/default/default.module#DefaultModule',
+        path: 'login',
+        loadChildren: () => import('../../modules/default/default.module').then(m => m.DefaultModule)
       },
        {
-        path: 'register', loadChildren: '../../modules/default/default.module#DefaultModule',
+        path: 'register',
+        loadChildren: () => import('../../modules/default/default.module').then(m => m.DefaultModule)
       },
     ]
   }
