@@ -4,16 +4,16 @@ import { FormsModule } from '@angular/forms';
 import { ReactiveFormsModule } from '@angular/forms';
 import { MailComposeComponent } from './mail-compose.component';
 import { Routes, RouterModule } from '@angular/router';
-import { AceEditorModule } from 'ng2-ace-editor';
-import { NgSelectModule } from '@ng-select/ng-select';
+// import { AceEditorModule } from 'ng2-ace-editor';
+// import { NgSelectModule } from '@ng-select/ng-select';
 @NgModule({
   imports: [
     CommonModule,
     FormsModule,
     ReactiveFormsModule,
     RouterModule,
-    AceEditorModule,
-    NgSelectModule
+    // AceEditorModule,
+    // NgSelectModule
   ],
   declarations: [
     MailComposeComponent,
