@@ -5,7 +5,7 @@ import { ReactiveFormsModule } from '@angular/forms';
 
 import { UserDetailComponent } from './user-detail.component';
 import { AlertModule } from '../../widgets/alert/alert.module';
-import { NgSelectModule } from '@ng-select/ng-select';
+// import { NgSelectModule } from '@ng-select/ng-select';
 
 @NgModule({
   imports: [
@@ -13,7 +13,7 @@ import { NgSelectModule } from '@ng-select/ng-select';
     FormsModule,
     ReactiveFormsModule,
     AlertModule,
-    NgSelectModule
+    // NgSelectModule
   ],
   declarations: [
     UserDetailComponent,
